fix commit 99d317e111267e3b18adcf287f9f70c7cb38a358 on 32 bits build

+++ Console/ImageHandler.h
@@ -59,7 +59,7 @@ public:
 		std::streamsize size = file.tellg();
 		file.seekg(0, std::ios::beg);
 
-		std::vector<char> content(size);
+		std::vector<char> content(static_cast<size_t>(size));
 		if( file.read(content.data(), size) )
 		{
 			return loadFromMemory(reinterpret_cast<const BYTE*>(content.data()), static_cast<DWORD>(content.size()));
